Models/current catalog sweep (#320)
* Model catalogs: drop deprecated ids, restore the Active GPT-4 tail

Swept every provider list, default, and doc against the vendors' own
deprecation pages on 2026-09-02.

Three ids in the picker could never have worked:

  - gpt-5.6-sol-pro / -terra-pro / -luna-pro are not model ids. "Pro" on
    the 5.6 family is reasoning.mode on the Responses API; OpenAI spells
    the replacement for gpt-5-pro-2025-10-06 as "gpt-5.6-sol
    (reasoning.mode: pro)".
  - gemini-3.1-pro and gemini-3-pro: the Pro line has no GA id. It ships
    as gemini-3.1-pro-preview, and gemini-3-pro-preview is already shut
    down. Same failure as the gemini-3.5-pro entry removed on 2026-08-10
    — an id written the way the marketing name reads.
  - deepseek-chat was retired 2026-07-24 and was also the provider's
    default, so every unconfigured DeepSeek call aimed at a dead id.

Live fallback defaults were worse: the vibecli TUI defaulted to
claude-3-opus-20240229 / gpt-4-turbo / gemini-pro / grok-beta, the
Bedrock config to claude-3-5-sonnet (retired 2025-10-28), and all six
candidates in hp_pick_provider were dead.

Removed: claude-sonnet-4-5, gpt-5, gpt-4-turbo, gemini-2.5-*,
gemini-2.0-*, grok-4.20, grok-3, deepseek-chat/-reasoner/-coder,
glm-4/-4v, glm-3-turbo, abab6.5*, llama-v3p3-70b, mixtral-8x*,
llama-3.3-70b-versatile, o1/o1-mini/o3-mini/o4-mini.

Added: claude-fable-5-1, gemini-3.8-flash, -3.7-flash, -3.1-flash-lite,
-3.1-pro-preview, grok-4.6. Azure and Copilot moved to the 5.6 line,
Fireworks to gpt-oss-120b / minimax-m3, SambaNova expanded 1 -> 6.

gpt-4o, gpt-4o-mini, gpt-4.1 and gpt-4.1-mini were cut as "superseded"
and are restored here, at the end of the OpenAI list. They are two
generations behind and OpenAI's guidance routes all four to 5.6, but
they are still callable and still listed Active — leaving ChatGPT on
2026-02-13 retired them from the consumer product, not from the API.
These lists drop ids that fail, not ids that are merely old. Their
deprecated neighbours (gpt-4.1-nano, gpt-4-turbo, gpt-3.5-turbo, the
o-line; API shutdown 2026-10-23) stay out, as does gpt-5.

Two bugs found alongside:

  - estimated_cost_usd matched no current model, so every Claude 5 /
    GPT-5.6 / Gemini 3 request fell through to _ => (0.0, 0.0) and was
    reported as free. Same hole in token_dashboard's pricing map. Both
    now carry list prices read from the vendors' pricing pages.
  - long_context_profiles carried Claude 3-era numbers (200K window,
    $15/$75) under Claude 5 names, so classify_context refused
    long-context routes the models support. Now 1M and $5/$25.

Verified: cargo check clean on vibe-ai, vibecli, vibe-coder; cargo test
-p vibecli -p vibe-ai --no-fail-fast passes 14,869; tsc --noEmit clean
for vibecoder and vibeaichat; vitest 1,722 pass. No panel was opened and
no live API call was made — every id is verified against vendor
documentation, not against a request.

Not touched: model_marketplace.rs still seeds gpt-4o / gemini-2.5-pro /
grok-3 with benchmark scores and Elo. Renaming those ids while keeping
GPT-4o's HumanEval number attached would be fabricated measurement, so
it needs real numbers or removal rather than a rename.


* feat: improve agent workflows and design tooling

---------

+++ vibecli/vibecli-cli/src/rl_executor.rs
@@ -198,7 +198,7 @@ impl PythonExecutor {
             );
             m.insert(
                 serde_yaml::Value::String("total_timesteps".into()),
-                serde_yaml::Value::Number((run.total_timesteps as i64).into()),
+                serde_yaml::Value::Number(run.total_timesteps.into()),
             );
             m.insert(
                 serde_yaml::Value::String("seed".into()),
